% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Computing multivariate Gaussian density for n p-dimensional observations}
\usage{
mvnpdf(X, mean, varcovM, Log = TRUE)
}
\arguments{
\item{X}{observation matrix, with n (columns) samples and p (rows) dimensions}

\item{mean}{p-dimensional mean vector of the MVN Gaussian density function}

\item{varcovM}{pxp dimensional Covariance matrix for the MVN Gaussian
density function}

\item{Log}{Logical argument, TRUE if the logarithm of the density function is
required. Default is TRUE}
}
\value{
a list containing the observation X as "Input Matrix" and their
corresponding (log) density as "MVN density"
}
\description{
Computing multivariate Gaussian density for n p-dimensional observations
}
\examples{
endeavr::mvnpdf(X = matrix(rep(1,6), 2, 3),
                 mean = c(1,2,3),
                 varcovM = diag(2),
                 Log = TRUE)
}
